+++ ThingsToDo/src/ThingsToDo.h
@@ -4,7 +4,7 @@
 //#include "resource.h"
 #include "ThingsToDo.h"
 //#include "MainWindow.h"
-#include "Main2Window.h"
+#include "MainWindow.h"
 
 int APIENTRY wWinMain(_In_ HINSTANCE hInstance,
     _In_opt_ HINSTANCE hPrevInstance,
